after adding possibility of updating student information by ID

+++ recogServer.py
@@ -1,8 +1,9 @@
 from base64 import b64encode
 import cv2
 import mysql
-from flask import Flask, render_template, request, Response, url_for, flash
+from flask import Flask, render_template, request, Response, url_for, flash, redirect
 from flask_paginate import get_page_parameter, Pagination
+
 import myRequests
 from camera import VideoCamera
 import test
@@ -10,13 +11,14 @@ import test
 app = Flask(__name__)
 ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg'}
 static_url_path = '/static'
+app.secret_key = 'key secret'
 
 db = test.db
 
 path = './images'
 img_path = []
 # for pagination
-limit = 5  # perPage
+limit = 6  # perPage
 
 rows = myRequests.studentInformation()
 total = len(rows)
@@ -46,35 +48,18 @@ def registeredSt():
 @app.route("/registerInfo", methods=['POST', 'GET'])
 def registerStudent():
     if request.method == 'POST':
-        if not request.form['id'] or not request.form['noms'] or not request.form['genre'] or not request.form['faculte'] or not request.form['departement'] or not request.form['promotion'] or not request.files['file']:
-            flash('Ples enter all the fields', 'error')
-        else:
-            Matricule = request.form["id"]
-            Noms = request.form["noms"]
-            Genre = request.form["genre"]
-            Faculte = request.form["faculte"]
-            Departement = request.form["departement"]
-            Promotion = request.form["promotion"]
-            Annee = request.form["annee"]
-            files = request.files["file"]
-            file = str(files)
-            # etudiantOb = test.Etudiant(matricule=Matricule, nom=Noms, genre=Genre, photo=file)
-            # studentObj = test.Student(Matricule, Noms, Genre, file)
+        Matricule = request.form["id"]
+        Noms = request.form["noms"]
+        Genre = request.form["genre"]
+        Faculte = request.form["faculte"]
+        Departement = request.form["departement"]
+        Promotion = request.form["promotion"]
+        Annee = request.form["annee"]
+        files = request.files["file"]
+        file = str(files)
         if request.form["submit"] == "SaveInfo":
             myRequests.RegisterStudent(Matricule, Noms, Genre, Faculte, Departement, Promotion, Annee, file)
-            # myRequests.RegisterFac(Faculte)
-            # myRequests.RegisterFincance(Matricule)
-            # myRequests.RegisterDepartem(Departement)
-            # myRequests.RegisterProm(Promotion)
-            # myRequests.RegisterAnnee(Matricule, Annee)
-            # myRequests.RegisterWps(Matricule)
-
-            # promotionOb = test.Promotion.query.all().first()
-
-            # db.session.add(studentObj)
-            # db.session.commit()
-
-            flash('Reccord was successfully added')
+            flash(' '+Noms+' was successfully added')
 
     # Pagination management
     page = request.args.get(get_page_parameter(), type=int, default=1)
@@ -94,9 +79,34 @@ def attendanceFinger():
     return render_template('attendance.html')
 
 
-@app.route("/studentInfo/<int:id>", methods=['POST', 'GET'])
+@app.route("/registerInfo/<int:id>", methods=['POST', 'GET'])
 def deleteStudent(id):
-    myRequests.deleteStudent(id)
+    if request.method == 'POST':
+        if request.form["submit"] == "delete":
+            myRequests.deleteStudent(id)
+            flash(' with ID '+id+' was deleted successfully in the database..:)')
+    # Pagination management
+    page = request.args.get(get_page_parameter(), type=int, default=1)
+    offset = page * limit - limit
+    studentsList = myRequests.paginationInfo(limit, offset)
+    pagination = Pagination(page=page, per_page=limit, total=total, css_framework='bootstrap4')
+    return render_template('studentInfo.html', student=studentsList, pagination=pagination, page=page)
+
+
+@app.route("/editInfo", methods=['POST', 'GET'])
+def editStudent():
+    if request.method == 'POST':
+        Matricule = request.form["idEdit"]
+        Noms = request.form["nomsEdit"]
+        Faculte = request.form["faculteEdit"]
+        Departement = request.form["departementEdit"]
+        Promotion = request.form["promotionEdit"]
+        Annee = request.form["anneeEdit"]
+        files = request.files["fileEdit"]
+        file = str(files)
+        if request.form['submit'] == "SaveEdit":
+            myRequests.editStudent(Matricule, Noms, Faculte, Departement, Promotion, Annee, file)
+            flash(' '+Noms+'\'s informations was updated successfully..:)')
     # Pagination management
     page = request.args.get(get_page_parameter(), type=int, default=1)
     offset = page * limit - limit
